requirements.txt
asgiref==3.2.7
Django==3.0.4
djangorestframework==3.11.0
gunicorn==20.0.4
pkg-resources==0.0.0
Pygments==2.6.1
pytz==2019.3
sqlparse==0.3.1
